Add string buffer util. Change default ecs tick elapsed to double.

+++ packages/ecs/include/ecs/entityecs/ECSExt.h
@@ -13,7 +13,7 @@
 #include <iostream>
 
 struct TickData {
-	float elapsedTime;
+	double elapsedTime;
 	float deltaTime;
 };
